dupl: read paths from command line arguments
Currently, dupl searches for clones in the current directory. There
is an option to search in another directory that can be specified
using the first argument.

This CL enables using other arguments as well. If the argument is
a directory it searches recursively all *.go files in that directory.
If it is a file it uses it regardless of the file extension.

The following commands are now equivalent:
	> dupl
	> dupl .
	> dupl $(find -name '*.go')
	> find -name '*.go' |dupl -files

Closes #7.